feat: InfluxDB 센서 평균값 조회 API 추가 및 소수점 정밀도 개선
feat: InfluxDB 센서 평균값 조회 API 추가 및 소수점 정밀도 개선

- 24시간 동안의 센서 데이터 평균값을 조회하는 API 엔드포인트를 추가합니다.
- InfluxDB에서 조회된 센서 값의 데이터 타입을 Double에서 BigDecimal로 변경하여 소수점 정밀도를 높였습니다.
- 센서 트렌드 조회 시 응답에 cultivationId, deviceEui, sensorType을 포함하도록 수정했습니다.
- 관련 테스트 코드를 수정된 로직에 맞게 업데이트했습니다.

+++ src/test/java/site/yesaido/cultivation_server/sensor/service/influx/InfluxAverageQueryTest.java
@@ -10,6 +10,7 @@ import site.yesaido.cultivation_server.config.InfluxProperties;
 import site.yesaido.cultivation_server.sensor.dto.response.influx.SensorTypeAverageResponse;
 import site.yesaido.cultivation_server.sensor.service.impl.InfluxServiceImpl;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -45,7 +46,7 @@ class InfluxAverageQueryTest {
                 service.findAverageByCultivationIdForLast24Hours(42L);
 
         assertThat(result).containsExactly(
-                new SensorTypeAverageResponse(42L, "TEMPERATURE", "C", 23.75)
+                new SensorTypeAverageResponse(42L, "TEMPERATURE", "C", BigDecimal.valueOf(23.75))
         );
 
         verify(queryApi).query(argThat((String query) ->
